Extract namespace tests
